requirements.txt
python-dotenv==1.0.0
requests==2.31.0 
pytest==7.4.0